+++ src/main/java/org/d2ab/iterator/ints/LimitingIntIterator.java
@@ -20,8 +20,7 @@ import java.util.NoSuchElementException;
 
 public class LimitingIntIterator extends UnaryIntIterator {
 	private final int limit;
-
-	private long count;
+	private int count;
 
 	public LimitingIntIterator(IntIterator iterator, int limit) {
 		super(iterator);
